requirements.txt
reflex
google-generativeai
python-dotenv